requirements.txt
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
chromedriver-py==115.0.5790.170
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
numpy==1.25.2
outcome==1.2.0
pandas==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
undetected-chromedriver==3.5.2
urllib3==1.26.16
websockets==11.0.3
wsproto==1.2.0
